chessGUI: Add event controller to gtk box
This is needed in gtk4 to handle signal events since eventboxes are
deprecated. In order to do this, labelBoard had to be made global since
we need to pass the outer widget, the eventbox, to the signal handler
button_pressed(), since the first parameter of the signal handler is not
the widget anymore but the gesture.