.github: Run on gtk container instead of installing it
Currently the dependency install step takes 14 seconds of the 19 second
build and test action. Hopefully using a docker image with dependencies
pre-installed will be faster.

I could not use the debian gtk image because it does not come with git
which is needed for cloning sub modules. The alpine image does not come
with make but installing it is super quick, so this should still be
faster than the default container.