chess.c: Refactor checkInput
It is easier to read the code when we exit early instead of having
several layers of if-nesting. In addition the tests them selves are
easier to read when '<' is used instead of '>='.

Added several tests that pass before and after refactoring. The tests
also fail if any of the return values are changed.