chess.c: Refactor makemove()
Remove nested if tests and return early instead, since it is easier to
read the code that way. For the same reason test for ' == 0' instead of
using '!' to negate.

Make the return statement of makemove be 1 for true and 0 for false to
match checkInput, and the C standard. Changing the other check functions
is still needed.