Apply clang-format style webkit
It is the closest to the style I prefer with only one exception, the
pieces enum where I want to keep the black and white pieces on their own
line.

Mostly the changes are to white-space, i.e. tabs to 4 spaces for indent.
I already changed to using spaces in university which is why some files
mix tabs and spaces.

It adds spaces some where, around math operators
for example and removes them elsewhere like around if parenthesis. No
big deal.

Also it adds the function opening { to a newline. I've been using this
style at work for a while now anyway.