Implement checkPawnPromotion
The function verifies if the promotion of a given pawn is legal. A
function for actually performing the promotion needs to be implemented.
Spliting it in two functions makes testing easier.

Removed inline from pieceOwnedByPlayer because gcc requires inline
functions to be defined in header files and I don't want to define
functions in the header files.